Continuando con el ABM cliente
Pantallas de consulta y alta

+++ AgenciaViajes/AgenciaDeViajesDB/dbo/Tables/Alojamiento_1.sql
@@ -1,9 +1,9 @@
 ﻿CREATE TABLE [dbo].[Alojamiento] (
-    [idAlojamiento]     NUMERIC (10)  NOT NULL,
-    [idTipoAlojamiento] NUMERIC (10)  NOT NULL,
+    [idAlojamiento]     INT  NOT NULL,
+    [idTipoAlojamiento] INT  NOT NULL,
     [domicilio]         NVARCHAR (50) NOT NULL,
     [nombre]            NVARCHAR (50) NOT NULL,
-    [telefono]          NUMERIC (13)  NOT NULL,
+    [telefono]          INT  NOT NULL,
     CONSTRAINT [PK_Alojamiento] PRIMARY KEY CLUSTERED ([idAlojamiento] ASC),
     CONSTRAINT [FK_Alojamiento_TipoAlojamiento] FOREIGN KEY ([idTipoAlojamiento]) REFERENCES [dbo].[TipoAlojamiento] ([idTipoAlojamiento])
 );
