+++ AgenciaViajes/AgenciaDeViajesDB/dbo/Tables/EmpresaColectivos_1.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [dbo].[EmpresaColectivos] (
-    [idEmpresaColectivos] NUMERIC (10)  NOT NULL,
+    [idEmpresaColectivos] INT  NOT NULL,
     [nombre]              NVARCHAR (50) NOT NULL,
-    [telefono]            NUMERIC (13)  NOT NULL,
+    [telefono]            INT  NOT NULL,
     [paginaWeb]           NVARCHAR (50) NULL,
     CONSTRAINT [PK_EmpresaColectivos] PRIMARY KEY CLUSTERED ([idEmpresaColectivos] ASC)
 );
