+++ AgenciaViajes/AgenciaDeViajesDB/dbo/Tables/EstadoCivil_1.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[EstadoCivil] (
-    [idEstadoCivil] NUMERIC (2)   NOT NULL,
+    [idEstadoCivil] INT   NOT NULL,
     [descripcion]   NVARCHAR (30) NOT NULL,
     CONSTRAINT [PK_EstadoCivil] PRIMARY KEY CLUSTERED ([idEstadoCivil] ASC)
 );
