+++ AgenciaViajes/AgenciaDeViajesDB/dbo/Tables/MotivoViaje_1.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[MotivoViaje] (
-    [idMotivoViaje] NUMERIC (2)   NOT NULL,
+    [idMotivoViaje] INT   NOT NULL,
     [descripcion]   NVARCHAR (30) NOT NULL,
     CONSTRAINT [PK_MotivoViaje] PRIMARY KEY CLUSTERED ([idMotivoViaje] ASC)
 );
