+++ AgenciaViajes/AgenciaDeViajesDB/dbo/Tables/SeguroViajero_1.sql
@@ -1,9 +1,9 @@
 ﻿CREATE TABLE [dbo].[SeguroViajero] (
-    [idSeguroViajero]   NUMERIC (10)  NOT NULL,
-    [comision]          NUMERIC (2)   NULL,
+    [idSeguroViajero]   INT  NOT NULL,
+    [comision]          INT   NULL,
     [monto]             FLOAT (53)    NULL,
-    [tipoSeguroViajero] NUMERIC (2)   NOT NULL,
-    [numeroCompra]      NUMERIC (10)  NULL,
+    [tipoSeguroViajero] INT   NOT NULL,
+    [numeroCompra]      INT  NULL,
     [descripcion]       NVARCHAR (50) NOT NULL,
     CONSTRAINT [PK_SeguroViajero] PRIMARY KEY CLUSTERED ([idSeguroViajero] ASC),
     CONSTRAINT [FK_SeguroViajero_Compra] FOREIGN KEY ([numeroCompra]) REFERENCES [dbo].[Compra] ([idCompra]),
