+++ AgenciaViajes/AgenciaDeViajesDB/dbo/Tables/TipoAlojamiento_1.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[TipoAlojamiento] (
-    [idTipoAlojamiento] NUMERIC (10)  NOT NULL,
+    [idTipoAlojamiento] INT  NOT NULL,
     [descripcion]       NVARCHAR (50) NOT NULL,
     CONSTRAINT [PK_TipoAlojamiento] PRIMARY KEY CLUSTERED ([idTipoAlojamiento] ASC)
 );
