+++ AgenciaViajes/AgenciaDeViajesDB/dbo/Tables/TipoDestino_1.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[TipoDestino] (
-    [idTipoDestino] NUMERIC (10)  NOT NULL,
+    [idTipoDestino] INT  NOT NULL,
     [descripcion]   NVARCHAR (50) NOT NULL,
     CONSTRAINT [PK_TipoDestino] PRIMARY KEY CLUSTERED ([idTipoDestino] ASC)
 );
