+++ AgenciaViajes/AgenciaDeViajesDB/dbo/Tables/TipoDocumento_1.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[TipoDocumento] (
-    [idTipoDocumento] NUMERIC (2)   NOT NULL,
+    [idTipoDocumento] INT   NOT NULL,
     [descripcion]     NVARCHAR (30) NOT NULL,
     CONSTRAINT [PK_tipoDocumento] PRIMARY KEY CLUSTERED ([idTipoDocumento] ASC)
 );
