+++ AgenciaViajes/AgenciaDeViajesDB/dbo/Tables/TipoSeguroViajero_1.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[TipoSeguroViajero] (
-    [tipoSeguroViajero] NUMERIC (2)   NOT NULL,
+    [tipoSeguroViajero] INT   NOT NULL,
     [descripcion]       NVARCHAR (30) NOT NULL,
     CONSTRAINT [PK_TipoSeguroViajero] PRIMARY KEY CLUSTERED ([tipoSeguroViajero] ASC)
 );
